+++ api-processor/src/test/kotlin/io/bkbn/lerasium/api/processor/KtorProcessorProviderTest.kt
@@ -94,8 +94,10 @@ class KtorProcessorProviderTest : DescribeSpec({
         import io.bkbn.kompendium.core.Notarized.notarizedGet
         import io.bkbn.kompendium.core.Notarized.notarizedPost
         import io.bkbn.kompendium.core.Notarized.notarizedPut
+        import io.bkbn.lerasium.generated.api.UserToC.countAllUser
         import io.bkbn.lerasium.generated.api.UserToC.createUser
         import io.bkbn.lerasium.generated.api.UserToC.deleteUser
+        import io.bkbn.lerasium.generated.api.UserToC.getAllUser
         import io.bkbn.lerasium.generated.api.UserToC.getUser
         import io.bkbn.lerasium.generated.api.UserToC.updateUser
         import io.bkbn.lerasium.generated.entity.UserDao
@@ -118,6 +120,12 @@ class KtorProcessorProviderTest : DescribeSpec({
                 val result = dao.create(request)
                 call.respond(result)
               }
+              notarizedGet(getAllUser) {
+                val chunk = call.parameters["chunk"]?.toInt() ?: 100
+                val offset = call.parameters["offset"]?.toInt() ?: 0
+                val result = dao.getAll(chunk, offset)
+                call.respond(result)
+              }
               route("/{id}") {
                 notarizedGet(getUser) {
                   val id = UUID.fromString(call.parameters["id"])
@@ -134,6 +142,12 @@ class KtorProcessorProviderTest : DescribeSpec({
                   val id = UUID.fromString(call.parameters["id"])
                   dao.delete(id)
                   call.respond(HttpStatusCode.NoContent)
+                }
+              }
+              route("/count") {
+                notarizedGet(countAllUser) {
+                  val result = dao.countAll()
+                  call.respond(result)
                 }
               }
             }
@@ -167,11 +181,14 @@ class KtorProcessorProviderTest : DescribeSpec({
         import io.bkbn.kompendium.core.metadata.method.PostInfo
         import io.bkbn.kompendium.core.metadata.method.PutInfo
         import io.bkbn.lerasium.api.model.GetByIdParams
+        import io.bkbn.lerasium.api.model.PaginatedQuery
+        import io.bkbn.lerasium.core.model.CountResponse
         import io.bkbn.lerasium.generated.models.UserCreateRequest
         import io.bkbn.lerasium.generated.models.UserResponse
         import io.bkbn.lerasium.generated.models.UserUpdateRequest
         import io.ktor.http.HttpStatusCode
         import kotlin.Unit
+        import kotlin.collections.List
 
         public object UserToC {
           public val createUser: PostInfo<Unit, UserCreateRequest, UserResponse> =
@@ -184,6 +201,17 @@ class KtorProcessorProviderTest : DescribeSpec({
             responseInfo = ResponseInfo(
               status = HttpStatusCode.Created,
               description = "The User was retrieved successfully"
+            ),
+            tags = setOf("User")
+          )
+
+
+          public val countAllUser: GetInfo<Unit, CountResponse> = GetInfo<Unit, CountResponse>(
+            summary = "Count User",
+            description = "Counts total User entities",
+            responseInfo = ResponseInfo(
+              status = HttpStatusCode.OK,
+              description = "Successfully retrieved the total User entity count"
             ),
             tags = setOf("User")
           )
@@ -221,6 +249,19 @@ class KtorProcessorProviderTest : DescribeSpec({
             responseInfo = ResponseInfo(
               status = HttpStatusCode.NoContent,
               description = "Successfully deleted User"
+            ),
+            tags = setOf("User")
+          )
+
+
+          public val getAllUser: GetInfo<PaginatedQuery, List<UserResponse>> =
+              GetInfo<PaginatedQuery, List<UserResponse>>(
+            summary = "Get All User",
+            description =
+                "Retrieves a collection of User entities, broken up by specified chunk and offset",
+            responseInfo = ResponseInfo(
+              status = HttpStatusCode.OK,
+              description = "Successfully retrieved the collection of User entities"
             ),
             tags = setOf("User")
           )