+++ api/src/main/kotlin/io/bkbn/lerasium/api/model/PaginatedQuery.kt
@@ -1,0 +1,11 @@
+package io.bkbn.lerasium.api.model
+
+import io.bkbn.kompendium.annotations.Param
+import io.bkbn.kompendium.annotations.ParamType
+
+data class PaginatedQuery(
+  @Param(ParamType.QUERY)
+  val chunk: Int = 100,
+  @Param(ParamType.QUERY)
+  val offset: Int = 0
+)