+++ mongo-processor/src/main/kotlin/io/bkbn/lerasium/mongo/processor/DaoVisitor.kt
@@ -14,15 +14,18 @@ import com.squareup.kotlinpoet.FileSpec
 import com.squareup.kotlinpoet.FunSpec
 import com.squareup.kotlinpoet.KModifier
 import com.squareup.kotlinpoet.MemberName
+import com.squareup.kotlinpoet.ParameterSpec
 import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
 import com.squareup.kotlinpoet.PropertySpec
 import com.squareup.kotlinpoet.TypeSpec
+import com.squareup.kotlinpoet.asClassName
 import com.squareup.kotlinpoet.asTypeName
 import com.squareup.kotlinpoet.ksp.KotlinPoetKspPreview
 import com.squareup.kotlinpoet.ksp.addOriginatingKSFile
 import com.squareup.kotlinpoet.ksp.toTypeName
 import io.bkbn.lerasium.core.Domain
 import io.bkbn.lerasium.core.dao.Dao
+import io.bkbn.lerasium.core.model.CountResponse
 import io.bkbn.lerasium.mongo.Unique
 import io.bkbn.lerasium.utils.KotlinPoetUtils.addControlFlow
 import io.bkbn.lerasium.utils.KotlinPoetUtils.addObjectInstantiation
@@ -88,6 +91,8 @@ class DaoVisitor(private val fileBuilder: FileSpec.Builder, private val logger: 
       addReadFunction(rc)
       addUpdateFunction(cd, urc, rc)
       addDeleteFunction()
+      addCountAllFunction()
+      addGetAllFunction(rc)
     }.build())
   }
 
@@ -203,6 +208,32 @@ class DaoVisitor(private val fileBuilder: FileSpec.Builder, private val logger: 
       addParameter("id", UUID::class)
       addCode(CodeBlock.builder().apply {
         addStatement("collection.%M(id)", DeleteOneById)
+      }.build())
+    }.build())
+  }
+
+  private fun TypeSpec.Builder.addCountAllFunction() {
+    addFunction(FunSpec.builder("countAll").apply {
+      addModifiers(KModifier.OVERRIDE)
+      returns(CountResponse::class)
+      addCode(CodeBlock.builder().apply {
+        addStatement("val count = collection.countDocuments()")
+        addStatement("return %T(count)", CountResponse::class)
+      }.build())
+    }.build())
+  }
+
+  private fun TypeSpec.Builder.addGetAllFunction(responseClass: ClassName) {
+    addFunction(FunSpec.builder("getAll").apply {
+      addModifiers(KModifier.OVERRIDE)
+      returns(List::class.asClassName().parameterizedBy(responseClass))
+      addParameter(ParameterSpec.builder("chunk", Int::class).build())
+      addParameter(ParameterSpec.builder("offset", Int::class).build())
+      addCode(CodeBlock.builder().apply {
+        addStatement("val entities = collection.find().skip(chunk * offset).limit(chunk)")
+        addControlFlow("return entities.toList().map { entity ->") {
+          addStatement("entity.toResponse()")
+        }
       }.build())
     }.build())
   }