+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Centralize test utils in util fixture
 - Omit sensitive fields from response payload
 - Unique constraint for document and table fields
+- Count All functionality
+- Get All functionality
 
 ### Changed